Main sketch: Fix the line that calculates the timing budget so that it works on processors with 16-bit integers.

+++ vl53l1x-st-api/vl53l1x-st-api.ino
@@ -94,7 +94,7 @@ void setup()
   status = VL53L1_DataInit(Dev);
   status = VL53L1_StaticInit(Dev);
   status = VL53L1_SetDistanceMode(Dev, VL53L1_DISTANCEMODE_LONG);
-  status = VL53L1_SetMeasurementTimingBudgetMicroSeconds(Dev, MEASUREMENT_BUDGET_MS * 1000);
+  status = VL53L1_SetMeasurementTimingBudgetMicroSeconds(Dev, (uint32_t)MEASUREMENT_BUDGET_MS * 1000);
   status = VL53L1_SetInterMeasurementPeriodMilliSeconds(Dev, INTER_MEASUREMENT_PERIOD_MS);
   status = VL53L1_StartMeasurement(Dev);
